UpdateCIScripts

Approved-by: Steven Hofmeyr

+++ ci/ci_asm_qual_test-full.sh
@@ -13,26 +13,44 @@ if [ ! -f "$refs" ]; then
     wget https://portal.nersc.gov/project/hipmer/MetaHipMer_datasets_12_2019/ArcticSynth/samples/${refs}.gz
     gunzip ${refs}.gz &
 fi
-reads=arctic_sample_0.fq
-if [ ! -f "$reads" ]; then
+all_reads=
+for i in $(seq 0 11)
+do
+  reads=arctic_sample_${i}.fq
+  if [ ! -f "$reads" ]; then
     rm -f ${reads}.gz
     wget https://portal.nersc.gov/project/hipmer/MetaHipMer_datasets_12_2019/ArcticSynth/samples/${reads}.gz
     gunzip ${reads}.gz &
-fi 
+  fi
+  if [ -n "$all_reads" ]
+  then
+    all_reads="${all_reads},${reads}"
+  else
+    all_reads="$reads"
+  fi
+done
 
 wait
 
+reads="${all_reads}"
+
 wd=`pwd`
-test_dir=$wd/test-arctic-sample0
-rm -rf $test_dir
-$mhm2_install_dir/bin/mhm2.py $@ -r $reads -o $test_dir --checkpoint=no --post-asm-align --post-asm-abd
+test_dir=$wd/test-arctic-samples
+if [ "$@" != "${@/--restart/}" ]
+then
+  rm -rf $test_dir
+else
+  echo "Restarting in $test_dir"
+fi
+$mhm2_install_dir/bin/mhm2.py $@ -r $reads -o $test_dir --force-bloom=yes --checkpoint=no --post-asm-align --post-asm-abd
 status=$?
 if [ $status -ne 0 ]
 then
   echo "MHM2 failed! - $status"
   exit $status
 fi
-$mhm2_install_dir/bin/check_asm_quality.py --asm-dir $test_dir --expected-quals $mhm2_install_dir/share/good-arctic-sample0.txt --refs $wd/$refs
+$mhm2_install_dir/bin/check_asm_quality.py --asm-dir $test_dir --expected-quals $mhm2_install_dir/share/good-arcticsynth.txt --refs $wd/$refs 2>&1 \
+   | tee $test_dir/check_asm_quality_test.log
 status=$?
 if [ $status -ne 0 ]
 then
